requirements.txt
requests==2.11.1
stitchstream-python==0.2.0
